Move raising fd limit to utils
